Read historical data by location

+++ src/main/java/com/java/course/model/Coordinates.java
@@ -5,19 +5,19 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class Coordinates {
 
-    private double latitude;
-    private double longitude;
+    private float latitude;
+    private float longitude;
 
-    public Coordinates(double latitude, double longitude) {
+    public Coordinates(float latitude, float longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
 
-    public double getLatitude() {
+    public float getLatitude() {
         return latitude;
     }
 
-    public double getLongitude() {
+    public float getLongitude() {
         return longitude;
     }
 }
